requirements.txt
google-genai>=0.3.0
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.9.0
requests>=2.32.0
